move flix types to separate file to avoid code size limits
