+++ packages/core/src/index.ts
@@ -10,3 +10,11 @@ export * from "./flow-storage.service";
 export * from "./flow-indexer.service";
 export * from "./flow-interactions.service";
 export * from "./flow-flix.service";
+export { FlowFlixServiceConfig } from "./flix-v1";
+export { FlixI18nMessage } from "./flix-v1";
+export { FlixMessages } from "./flix-v1";
+export { FlixDependencyOnNetwork } from "./flix-v1";
+export { FlixDependency } from "./flix-v1";
+export { FlixArgument } from "./flix-v1";
+export { FlixAuditor } from "./flix-v1";
+export { FlixTemplateV1 } from "./flix-v1";
